drawer navigation at top for login view and logout
